rpg-backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
python-dotenv
django-cors-headers
